Hero map overlay, event cards, dark mode fixes
