+++ scripts/get-github-info/last-success.isodate
@@ -1,1 +1,1 @@
-2026-03-27T00:00:00.567Z
+2026-03-27T02:25:00.665Z
