+++ src/app/day/2026/components/why-attend-section.tsx
@@ -4,28 +4,28 @@ export function WhyAttendSection() {
       <h3 className="typography-h2 mb-12">Why attend GraphQL Day?</h3>
       <div className="grid gap-8 md:grid-cols-2 lg:grid-cols-3">
         <Card
-          title="Technical Deep Dives"
-          description="Learn advanced patterns, performance optimization techniques, and architectural decisions from teams running GraphQL at scale."
+          title="Real-World Talks"
+          description="Hear how teams actually use GraphQL — what worked, what didn't, and what they'd do differently."
         />
         <Card
-          title="Best Practices Exchange"
-          description="Share your learnings and challenges with fellow practitioners. Discover how others solve common GraphQL problems."
+          title="Cross-Community"
+          description="FOST brings together GraphQL, AsyncAPI, OpenAPI, and JSON Schema communities. Learn across boundaries."
         />
         <Card
-          title="Innovation Showcase"
-          description="Explore cutting-edge tools, upcoming features, and experimental approaches that are shaping GraphQL's future."
+          title="Open Source"
+          description="Meet library maintainers and core contributors. Ask questions, report bugs, discuss roadmaps."
         />
         <Card
-          title="Community Building"
-          description="Connect with the GraphQL community. Build relationships that extend beyond the conference."
+          title="Workshops"
+          description="Hands-on sessions to try new tools and techniques with guidance from the people who build them."
         />
         <Card
-          title="Hands-on Learning"
-          description="Interactive sessions where you can experiment with new tools and techniques in real-time."
+          title="Networking"
+          description="A focused single-day format means everyone's in the same room. Easy to meet people."
         />
         <Card
-          title="Q&A Sessions"
-          description="Direct access to library maintainers and core contributors. Get your specific questions answered."
+          title="Free for Speakers"
+          description="All speakers get a free conference ticket. Submit a talk and share what you've learned."
         />
       </div>
     </section>
