+++ src/app/day/events-map.tsx
@@ -21,6 +21,8 @@ import {
   MapColors,
 } from "@/app/(main)/community/events/map/map-colors"
 
+import { EVENTS } from "./events-data"
+
 const CELL_SIZE = 8
 const SQUARE_SIZE = 6
 const LAND_MASK_URL = new URL(
@@ -30,58 +32,6 @@ const LAND_MASK_URL = new URL(
 const ASPECT_RATIO = 1.65
 
 type ThemeVariant = keyof typeof MAP_COLORS
-
-export interface EventMapItem {
-  id: string
-  city: string
-  date: string
-  href: string
-  lon: number
-  lat: number
-}
-
-const EVENTS: EventMapItem[] = [
-  {
-    id: "singapore",
-    city: "Singapore",
-    date: "Apr 14-15",
-    href: "/day/2026/singapore",
-    lon: 103.8198,
-    lat: 1.3521,
-  },
-  {
-    id: "nyc",
-    city: "NYC",
-    date: "May 13-14",
-    href: "/day/2026/nyc",
-    lon: -74.006,
-    lat: 40.7128,
-  },
-  {
-    id: "amsterdam",
-    city: "Amsterdam",
-    date: "Jun 9-10 [TBC]",
-    href: "/day/2026/amsterdam",
-    lon: 4.9041,
-    lat: 52.3676,
-  },
-  {
-    id: "melbourne",
-    city: "Melbourne",
-    date: "Oct 28-29",
-    href: "/day/2026/melbourne",
-    lon: 144.9631,
-    lat: -37.8136,
-  },
-  {
-    id: "paris",
-    city: "Paris",
-    date: "Dec 1-3",
-    href: "/day/2026/paris",
-    lon: 2.3522,
-    lat: 48.8566,
-  },
-]
 
 const markerPoints: MarkerPoint[] = EVENTS.map(e => ({
   id: e.id,
